# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoffeeAndStudy API",
    "description": "API REST para gestionar lugares de estudio, usuarios, reseñas y sesiones de estudio",
    "contact": {
      "name": "Ibone",
      "email": "ibone23@example.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/places/{id}": {
      "get": {
        "tags": [
          "Places"
        ],
        "summary": "Obtener place por id",
        "description": "Devuelve un lugar concreto a partir de su identificador",
        "operationId": "getPlaceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Place encontrado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Place no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Places"
        ],
        "summary": "Actualizar place",
        "description": "Actualiza un lugar existente a partir de su id",
        "operationId": "updatePlace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Place actualizado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Place no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Places"
        ],
        "summary": "Eliminar place",
        "description": "Elimina un lugar a partir de su id",
        "operationId": "deletePlace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Place eliminado correctamente"
          },
          "404": {
            "description": "Place no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Obtener categoría por id",
        "description": "Devuelve una categoría concreta a partir de su identificador",
        "operationId": "getCategoryById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría encontrada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Categoría no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Actualizar categoría",
        "description": "Actualiza una categoría existente a partir de su id",
        "operationId": "updateCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Categoría no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Eliminar categoría",
        "description": "Elimina una categoría a partir de su id",
        "operationId": "deleteCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Categoría eliminada correctamente"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/places": {
      "get": {
        "tags": [
          "Places"
        ],
        "summary": "Obtener todos los places",
        "description": "Devuelve la lista de lugares, con posibilidad de filtrar por ciudad, wifi y activo",
        "operationId": "getAllPlaces",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "wifiAvailable",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de places obtenida correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de filtro inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Places"
        ],
        "summary": "Crear place",
        "description": "Crea un nuevo lugar de estudio",
        "operationId": "createPlace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Place creado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Obtener todas las categorías",
        "description": "Devuelve la lista de categorías, con posibilidad de filtrar por nombre, activo y prioridad",
        "operationId": "getAllCategories",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías obtenida correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de filtro inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Crear categoría",
        "description": "Crea una nueva categoría",
        "operationId": "createCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Categoría creada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0
          },
          "colorHex": {
            "type": "string",
            "minLength": 1,
            "pattern": "^#([A-Fa-f0-9]{6})$"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "active",
          "colorHex",
          "description",
          "name",
          "priority"
        ]
      },
...
      "PlaceRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "maximum": 90,
            "minimum": -90
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "maximum": 180,
            "minimum": -180
          },
          "city": {
            "type": "string"
          },
          "wifiAvailable": {
            "type": "boolean"
          },
          "averagePrice": {
            "type": "number",
            "format": "float"
          },
          "openingDate": {
            "type": "string",
            "format": "date"
          },
          "active": {
            "type": "boolean"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "categoryId",
          "description",
          "latitude",
          "longitude",
          "title"
        ]
      },
      "PlaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "categoryName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "city": {
            "type": "string"
          },
          "wifiAvailable": {
            "type": "boolean"
          },
          "averagePrice": {
            "type": "number",
            "format": "float"
          },
          "openingDate": {
            "type": "string",
            "format": "date"
          },
          "active": {
            "type": "boolean"
          }
        }
      }
    }
  }
}